fix type

+++ test/test_sync_subscriptions.cpp
@@ -1102,7 +1102,7 @@ TEST(Sync_MutableSubscriptionReleasesReadLock)
     Query query_a(read_tr->get_table("class_a"));
     query_a.greater_equal(fixture.bar_col, int64_t(1));
 
-    size_t num_versions_before_subscription = fixture.db->get_number_of_versions();
+    uint_fast64_t num_versions_before_subscription = fixture.db->get_number_of_versions();
     auto mut_subs = store->get_latest().make_mutable_copy();
     auto [it, inserted] = mut_subs.insert_or_assign("a sub", query_a);
     CHECK(inserted);
@@ -1116,7 +1116,7 @@ TEST(Sync_MutableSubscriptionReleasesReadLock)
         wt->commit();
         read_tr->advance_read(); // update our reader so that its old version can be cleaned up
     }
-    size_t num_versions_after_sync_writes = fixture.db->get_number_of_versions();
+    uint_fast64_t num_versions_after_sync_writes = fixture.db->get_number_of_versions();
 
     // check that the mut_subs in not keeping a transaction pinned
     CHECK_EQUAL(num_versions_after_sync_writes, num_versions_before_subscription);
